Implemented the log writer function
