+++ HACC_App/src/Log.txt
@@ -1,4 +1,4 @@
 			------LOG SYSTEM OF HACC------
 
-The light turned on by User
+The light turned on by User on Fri Sep 11 18:31:08 BDT 2020
 The light turned of by User
